fix(midway-mock): types structure
