+++ packages/midway-mock/src/interface.ts
@@ -1,7 +1,6 @@
-import { MockOption } from 'egg-mock';
+import { MockOption, MockApplication  } from 'egg-mock';
+import { ApplicationContext, IApplicationContext } from 'injection';
 
-
-export { MidwayMockApplication } from './app/extend/application';
 
 export interface MidwayApplicationOptions extends MockOption {
   baseDir?: string
@@ -11,5 +10,34 @@ export interface MidwayApplicationOptions extends MockOption {
   container?: any
   typescript?: boolean
   worker?: number
+}
+
+export interface MidwayMockApplication extends MockApplication {
+  applicationContext: ApplicationContext
+  pluginContext: IApplicationContext
+  appDir: string
+  baseDir: string
+  enablePlugins: any
+  getApplicationContext(): IApplicationContext
+  getPluginContext(): IApplicationContext
+  getPlugin(pluginName: string): any
+  getLogger(name?: string): any
+  getConfig(key?: string): any
+  _mockFn(
+    service: string,
+    methodName: string,
+    /** {Object|Function|Error} - mock you data */
+    fnOrData: any,
+  ): void
+
+  /**
+   * Mock class function
+   */
+  mockClassFunction(
+    this: MidwayMockApplication,
+    className: string,
+    methodName: string,
+    fnOrData: any,
+  ): any
 }
